+++ model/src/test/java/se/somath/publisher/includer/FileTreeIncluderTest.java
@@ -6,6 +6,7 @@ import se.somath.publisher.builder.FileTreeBuilder;
 import java.util.LinkedList;
 import java.util.List;
 
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
@@ -14,6 +15,7 @@ public class FileTreeIncluderTest {
     @Test
     public void shouldCallReadFileWithIncludeOnOneLine() {
         List<String> given = new LinkedList<String>();
+        boolean onlyDirs = false;
 
         given.add("<include-file-tree root=\"fileRoot\"/>");
 
@@ -23,8 +25,31 @@ public class FileTreeIncluderTest {
 
         includer.addIncludes(given);
 
-        verify(fileTreeBuilder).buildFileTree("fileRoot");
+        verify(fileTreeBuilder).buildFileTree("fileRoot", onlyDirs);
     }
 
+    @Test
+    public void shouldOnlyIncludeDirectories() {
+        List<String> given = new LinkedList<String>();
 
+        given.add("<include-file-tree root=\"model\"");
+        given.add("                   only-dirs=\"true\"/>");
+
+        FileTreeBuilder fileTreeBuilder = new FileTreeBuilder();
+        Includer includer = new Includer();
+        includer.setFileTreeBuilder(fileTreeBuilder);
+
+        List<String> actual = includer.addIncludes(given);
+
+        assertNoFiles(actual);
+
+    }
+
+    private void assertNoFiles(List<String> actual) {
+        for (String candidate : actual) {
+            if (candidate.endsWith(".java") || candidate.endsWith(".xml")) {
+                fail("Found a file: " + candidate + " and did not expect to find any file");
+            }
+        }
+    }
 }