+++ History.md
@@ -1,6 +1,6 @@
 ## [Git master] (https://github.com/tsundberg/publisher/compare/v1.0.4...master)
 -----------------------------
-* None yet
+* Added support for including the only directories in a file tree
 
 ## [v1.0.4] (https://github.com/tsundberg/publisher/compare/v1.0.3...v1.0.4)
 -----------------------------